app/requirements.txt
aiomysql==0.2.0
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
devtools==0.12.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.16.1
PyMySQL==1.1.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
